+++ quotegen/src/main/java/com/andmark/quotegen/service/UserService.java
@@ -3,6 +3,7 @@ package com.andmark.quotegen.service;
 import com.andmark.quotegen.domain.User;
 import com.andmark.quotegen.domain.enums.UserRole;
 import com.andmark.quotegen.dto.UserDTO;
+import org.springframework.transaction.annotation.Transactional;
 
 public interface UserService extends AbstractService<UserDTO>{
     boolean isRegistered(Long id);
@@ -14,6 +15,9 @@ public interface UserService extends AbstractService<UserDTO>{
     void changePassword(String username, String currentPassword, String newPassword);
 
     UserDTO findOneByUsertgId(Long usertgId);
+
+    @Transactional
+    void deleteByUsertgId(Long id);
 
     void deleteByUsername(String username);