+++ quotegen/src/main/resources/static/version-info/CHANGELOG.md
@@ -9,6 +9,8 @@
 ## [2.0.1] - 2024-01-14
 ### Fixed
 - added the ability to automatically create an ngrok tunnel in the event of a shutdown
+- readme command
+- signout command, delete by usertgid
 ## [2.0.0-RELEASE] - 2024-01-13
 ### Added
 - added clearing the database of books with DELETED status